+++ analytics_event/example/lib/example.g.dart
@@ -8,22 +8,26 @@ part of 'example.dart';
 
 // ignore_for_file: unnecessary_statements
 
+// useNullSafetySyntax: true
 class _$Events extends Events with AnalyticsEventStubsImpl {
-  _$Events([TrackAnalytics tracker]) {
+  _$Events([TrackAnalytics? tracker]) {
     tracker != null ? registerTracker(tracker) : null;
   }
 
   @override
-  void trackAppLaunch({String date}) =>
-      trackEvent('appLaunch', <String, dynamic>{'date': date});
+  void trackAppLaunch({required String date}) =>
+      trackEvent('appLaunch', <String, Object?>{'date': date});
   @override
   void trackExample(String myRequiredParameter,
           {String withDefault = 'test'}) =>
-      trackEvent('example', <String, dynamic>{
+      trackEvent('example', <String, Object?>{
         'myRequiredParameter': myRequiredParameter,
         'withDefault': withDefault
       });
   @override
   void trackItem(ItemAction action) => trackEvent(
-      'item', <String, dynamic>{'action': action?.toString()?.substring(11)});
+      'item', <String, Object?>{'action': action.toString().substring(11)});
+  @override
+  void trackNullableItem(ItemAction? action) => trackEvent('nullableItem',
+      <String, Object?>{'action': action?.toString().substring(11)});
 }